# packages/api-client-ts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/children": {
      "get": {
        "operationId": "Children_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChildDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The signed-in parent's children.",
        "tags": [
          "children"
        ]
      },
      "post": {
        "operationId": "Children_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChildDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creates a child account and records the parent's consent.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/rules": {
      "get": {
        "operationId": "Children_rules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildRulesDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "What the \"add a child\" form may offer: birth years and avatars.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/nickname-suggestions": {
      "get": {
        "operationId": "Children_suggestions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NicknameSuggestionsDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Safe nickname ideas, e.g. \"SwiftFalcon27\".",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{id}": {
      "get": {
        "operationId": "Children_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "children"
        ]
      },
      "patch": {
        "operationId": "Children_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChildDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change nickname, avatar, language or location.",
        "tags": [
          "children"
        ]
      },
      "delete": {
        "operationId": "Children_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteChildDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes the child's account and personal data. Confirm with the nickname.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{id}/consents": {
      "put": {
        "operationId": "Children_setConsents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildConsentsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Switch public leaderboards and the public portfolio on or off.",
        "tags": [
          "children"
        ]
      },
      "get": {
        "operationId": "Children_consents",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsentRecordDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every consent the parent has given or withdrawn for this child.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{id}/portfolio": {
      "get": {
        "operationId": "Children_portfolio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Language for titles and texts, e.g. \"ar\". Falls back to English for anything untranslated.",
            "schema": {
              "pattern": "^[a-z]{2}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildPortfolioDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The child's shipped projects, and the share link if the parent made one.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{id}/portfolio/share-link": {
      "post": {
        "operationId": "Children_createShareLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioShareDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A new share link for the portfolio (the old one stops working). Needs \"Public projects\".",
        "tags": [
          "children"
        ]
      },
      "delete": {
        "operationId": "Children_removeShareLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stops the share link working.",
        "tags": [
          "children"
        ]
      }
    },
    "/v1/children/{id}/password": {
      "post": {
        "operationId": "Children_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetChildPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in"
          },
          "403": {
            "description": "Signed in, but not allowed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sets a new password for the child and signs them out everywhere.",
        "tags": [
          "children"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kids Coding Platform API",
    "description": "One API for the student and parent web app, the admin panel, the marketing site and the Flutter app. Every route checks the caller’s role.",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChildConsentsDto": {
        "type": "object",
        "properties": {
          "publicLeaderboards": {
            "type": "boolean",
            "description": "Show the child (nickname and avatar only) on city, country and global leaderboards."
          },
          "publicPortfolio": {
            "type": "boolean",
            "description": "Let anyone with the link see the child's finished projects."
          }
        },
        "required": [
          "publicLeaderboards",
          "publicPortfolio"
        ]
      },
      "ChildDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "description": "The child's login name."
          },
          "nickname": {
            "type": "string"
          },
          "avatarKey": {
            "type": "string"
          },
          "birthYear": {
            "type": "number"
          },
          "languageCode": {
            "type": "string"
          },
          "countryCode": {
            "type": "string",
            "nullable": true
          },
          "regionId": {
            "type": "string",
            "nullable": true
          },
          "cityId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "SUSPENDED",
              "DELETED"
            ],
            "type": "string"
          },
          "consents": {
            "$ref": "#/components/schemas/ChildConsentsDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "lessonsCompleted": {
            "type": "number",
            "description": "Lessons the child has completed."
          },
          "premiumUntil": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "When today's premium ends (a plan renews on its own); null without premium."
          },
          "premiumSource": {
            "nullable": true,
            "description": "Where premium comes from: the family's plan, our team, or the free trial.",
            "enum": [
              "subscription",
              "grant",
              "trial"
            ],
            "type": "string"
          },
          "trialEndsAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "The child's free trial (over or not)."
          },
          "xpTotal": {
            "type": "number",
            "description": "All XP earned so far."
          },
          "level": {
            "type": "number"
          },
          "streak": {
            "type": "number",
            "description": "Days in a row with the daily goal met (0 once a day is missed)."
          },
          "badges": {
            "type": "number",
            "description": "Badges earned."
          }
        },
        "required": [
          "id",
          "username",
          "nickname",
          "avatarKey",
          "birthYear",
          "languageCode",
          "countryCode",
          "regionId",
          "cityId",
          "status",
          "consents",
          "createdAt",
          "lastLoginAt",
          "lessonsCompleted",
          "premiumUntil",
          "premiumSource",
          "trialEndsAt",
          "xpTotal",
          "level",
          "streak",
          "badges"
        ]
      },
      "ChildRulesDto": {
        "type": "object",
        "properties": {
          "birthYears": {
            "description": "Birth years a parent may choose right now.",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "avatarKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "under13Open": {
            "type": "boolean",
            "description": "Whether accounts for children under 13 are open."
          }
        },
        "required": [
          "birthYears",
          "avatarKeys",
          "under13Open"
        ]
      },
      "NicknameSuggestionsDto": {
        "type": "object",
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "suggestions"
        ]
      },
      "CreateChildDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "Shown to other children. Letters, digits and _; never the child's real name."
          },
          "avatarKey": {
            "type": "string",
            "enum": [
              "rocket",
              "star",
              "bolt",
              "planet",
              "robot",
              "leaf",
              "moon",
              "sun",
              "cube",
              "gamepad",
              "music",
              "code"
            ]
          },
          "birthYear": {
            "type": "number",
            "description": "Only the year is stored, never the full date of birth."
          },
          "countryCode": {
            "type": "string",
            "pattern": "^[A-Z]{2}$"
          },
          "regionId": {
            "type": "string",
            "format": "uuid"
          },
          "cityId": {
            "type": "string",
            "format": "uuid"
          },
          "languageCode": {
            "type": "string",
            "description": "Language the child learns in, e.g. \"ur\".",
            "pattern": "^[a-z]{2}$"
          },
          "password": {
            "type": "string",
            "description": "Chosen by the parent; the child logs in with it.",
            "minLength": 8,
            "maxLength": 128
          },
          "consents": {
            "$ref": "#/components/schemas/ChildConsentsDto"
          }
        },
        "required": [
          "nickname",
          "avatarKey",
          "birthYear",
          "countryCode",
          "languageCode",
          "password",
          "consents"
        ]
      },
      "UpdateChildDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          },
          "avatarKey": {
            "type": "string",
            "enum": [
              "rocket",
              "star",
              "bolt",
              "planet",
              "robot",
              "leaf",
              "moon",
              "sun",
              "cube",
              "gamepad",
              "music",
              "code"
            ]
          },
          "languageCode": {
            "type": "string",
            "pattern": "^[a-z]{2}$"
          },
          "regionId": {
            "type": "string",
            "format": "uuid"
          },
          "cityId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ConsentRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "ACCOUNT",
              "PUBLIC_LEADERBOARDS",
              "PUBLIC_PORTFOLIO",
              "HUB_WORK",
              "EARNINGS"
            ]
          },
          "policyVersion": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "grantedAt": {
            "format": "date-time",
            "type": "string"
          },
          "revokedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "policyVersion",
          "method",
          "grantedAt",
          "revokedAt"
        ]
      },
      "PortfolioShareDto": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean",
            "description": "True while \"Public projects\" is on: only then can the portfolio be shared."
          },
          "token": {
            "type": "string",
            "nullable": true,
            "description": "The secret part of the share link, or null when there is no link."
          }
        },
        "required": [
          "allowed",
          "token"
        ]
      },
      "CodeFilesDto": {
        "type": "object",
        "properties": {
          "html": {
            "type": "string",
            "maxLength": 20000
          },
          "css": {
            "type": "string",
            "maxLength": 20000
          },
          "js": {
            "type": "string",
            "maxLength": 20000
          },
          "py": {
            "type": "string",
            "description": "Python lessons: the program (runs with Pyodide in the sandbox).",
            "maxLength": 20000
          }
        }
      },
      "PortfolioItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "description": "The project's title, in the requested language."
          },
          "moduleTitle": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string"
          },
          "files": {
            "description": "The shipped files, to show in the sandbox.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CodeFilesDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "moduleTitle",
          "version",
          "publishedAt",
          "files"
        ]
      },
      "ChildPortfolioDto": {
        "type": "object",
        "properties": {
          "share": {
            "$ref": "#/components/schemas/PortfolioShareDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortfolioItemDto"
            }
          }
        },
        "required": [
          "share",
          "items"
        ]
      },
      "ResetChildPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "password"
        ]
      },
      "DeleteChildDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "description": "Type the child's nickname to confirm.",
            "maxLength": 40
          }
        },
        "required": [
          "nickname"
        ]
      },
...
    }
  }
}